+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 - New #597, #608: Add debug collector for `yiisoft/yii-debug` (@xepozz, @vjik)
 - New #610: Add `$escape` parameter to methods `Result::getAttributeErrorMessagesIndexedByPath()` and
   `Result::getErrorMessagesIndexedByPath()` that allow change or disable symbol which will be escaped in value path
-  elements. 
+  elements (@vjik)
 - Bug #612: Disable escaping of asterisk char in value path that returns by `Error::getValuePath(true)` (@vjik)
 
 ## 1.1.0 April 06, 2023