+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 - New #610: Add `$escape` parameter to methods `Result::getAttributeErrorMessagesIndexedByPath()` and
   `Result::getErrorMessagesIndexedByPath()` that allow change or disable symbol which will be escaped in value path
   elements (@vjik)
-- Bug #612: Disable escaping of asterisk char in value path that returns by `Error::getValuePath(true)` (@vjik)
+- Bug #612: Disable escaping of asterisk char in value path returned by `Error::getValuePath(true)` (@vjik)
 
 ## 1.1.0 April 06, 2023