+++ CHANGELOG.md
@@ -7,6 +7,7 @@
   `Result::getErrorMessagesIndexedByPath()` that allow change or disable symbol which will be escaped in value path
   elements (@vjik)
 - Bug #612: Disable escaping of asterisk char in value path returned by `Error::getValuePath(true)` (@vjik)
+- New #617: Add `OneOf` rule (@arogachev)
 
 ## 1.1.0 April 06, 2023